Move rendering folder to client and rename to renderer
