+++ common/src/main/kotlin/io/github/priestoffern/vs_ship_assembler/client/renderer/RenderingData.kt
@@ -1,4 +1,4 @@
-package io.github.priestoffern.vs_ship_assembler.rendering
+package io.github.priestoffern.vs_ship_assembler.client.renderer
 
 import com.mojang.blaze3d.vertex.PoseStack
 import net.minecraft.client.Camera
